requirements.txt
python-dotenv
requests
selectolax
loguru